containers/seldon-score/requirements.txt
Flask==1.1.1
gunicorn==20.0.4
numpy
tensorflow==2.0.0-alpha0
Pillow
requests
